Move FloatingText to UI namespace
