+++ Assets/Scripts/Skills/Fishing/Core/FisherController.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using UnityEngine.EventSystems;
 using Inventory;
 using Player;
-using Skills.Mining;
+using UI;
 
 namespace Skills.Fishing
 {
