+++ Assets/Scripts/UI/FloatingText.cs.meta
